Env map reactivity fixes
